requirements.txt
prometheus-client==0.19.0
mysql-connector-python==8.3.0